Split GitHub Actions history into focused files
